requirements.txt
Flask==2.3.3
mysql-connector-python==8.1.0
python-dotenv==1.0.0
Werkzeug==2.3.7
itsdangerous==2.1.2
Jinja2==3.1.6
MarkupSafe==3.0.3
requests
unidecode
google-genai